Adjusted some UI for Headers, Slack image, Reset password.
