+++ src/js/components/Person/EmailOfficialManager.jsx
@@ -1,4 +1,5 @@
-import { TextField } from '@mui/material';
+import { Alert, TextField } from '@mui/material';
+import { ErrorOutline } from '@mui/icons-material';
 import { withStyles } from '@mui/styles';
 import PropTypes from 'prop-types';
 import React, { useEffect, useState } from 'react';
@@ -27,12 +28,16 @@ const EmailOfficialManager = (
   const { viewerAccessRights } = apiDataCache;
 
   const [activePerson, setActivePerson] = useState(getAppContextValue('profileDrawerPerson'));
+  const [emailError, setEmailError] = useState('');
   const [emailOfficialNotValidDomain, setEmailOfficialNotValidDomain] = useState(false);
   const [emailOfficialVerifiedLocally, setEmailOfficialVerifiedLocally] = useState(false);
   const [emailOfficialVerifiedToNotExist, setEmailOfficialVerifiedToNotExist] = useState(false);
   const [isEmailOfficialEditModeOn, setIsEmailOfficialEditModeOn] = useState(false);
   const [newAccountNotification, setNewAccountNotification] = useState('');
   const [newAccountNotificationCopied, setNewAccountNotificationCopied] = useState(false);
+  const [newPasswordNotification, setNewPasswordNotification] = useState('');
+  const [newPasswordNotificationCopied, setNewPasswordNotificationCopied] = useState(false);
+  const [showNewPasswordNotification, setShowNewPasswordNotification] = useState(false);
   const [requiredVariablesMissingMessage, setRequiredVariablesMissingMessage] = useState('');
   const [resultsText, setResultsText] = useState('');
   const [showNewAccountNotification, setShowNewAccountNotification] = useState(false);
@@ -57,6 +62,13 @@ const EmailOfficialManager = (
       setEmailOfficialInParent(savedEmailOfficial);
     }
     setIsEmailOfficialEditModeOn(false);
+  };
+
+  const newPasswordNotificationOnCopy = () => {
+    setNewPasswordNotificationCopied(true);
+    setTimeout(() => {
+      setNewPasswordNotificationCopied(false);
+    }, 1500);
   };
 
   const reformatPhoneNumberToGooglePattern = (phoneNumber) => {
@@ -114,7 +126,32 @@ Please note that we prefer that you keep your Slack account connected to your pe
   };
 
   const resetPassword = async () => {
-    //
+    setEmailError('');
+    const { emailOfficial, emailPersonal, firstName } = activePerson;
+    if (!emailOfficial.includes('wevoteeducation.org') && !emailPersonal.includes('wevoteeducation.org')) {
+      setEmailError(`"${firstName}" does not have a wevoteeducation.org email as either their official (${emailOfficial}) or private (${emailOfficial}) address`);
+      return;
+    }
+    const emailToChange = emailOfficial.includes('wevoteeducation.org') ? emailOfficial : emailPersonal;
+    const newPassword = generateRandomString(12);
+    // console.log(`deletewevoteuserpassword ${emailToChange} ${newPassword}`);
+    const data = await weConnectQueryFn('google-reset-user-password', {
+      primaryEmail: emailToChange,
+      newPassword,
+    }, METHOD.POST);
+    // console.log('reset wevote password', data);
+    if (data.success) {
+      setNewPasswordNotification(`Thank you for requesting that your password be reset. These are your new sign in credentials: \n https://team.wevote.org \n user: ${emailToChange} \n pass: ${newPassword}`);
+      setShowNewPasswordNotification(true);
+    } else {
+      // console.error('Failed to reset password:', data);
+      console.log(`Reset password API error: ${data.error}`);
+      if (data.error.startsWith('Missing required')) {
+        setEmailError(`'${emailToChange}' is not an existing valid @wevoteeducation.org email address.`);
+      } else {
+        setEmailError(`${data.error}, google api server code: ${data.errorCode}`);
+      }
+    }
   };
 
   const verifyEmail = async () => {
@@ -140,7 +177,7 @@ Please note that we prefer that you keep your Slack account connected to your pe
       } else {
         verifiedJustNow = !!(verificationData.creationTime);
         verifiedToNotExistJustNow = !(verificationData.creationTime);
-        setActivePerson(prev => ({
+        setActivePerson((prev) => ({
           ...prev,
           emailOfficialVerified: true,
         }));
@@ -215,7 +252,7 @@ Please note that we prefer that you keep your Slack account connected to your pe
   const showEditEmailLink = emailOfficialExistsInDbAndUnchanged && !isEmailOfficialEditModeOn;
   const showEmailVerified = emailOfficialExistsInDbAndUnchanged && (!activePerson.emailOfficialVerified && emailOfficialVerifiedLocally);
   const showSavedEmailOfficialVerified = emailOfficialExistsInDbAndUnchanged && (activePerson.emailOfficialVerified);
-  const showResetPassword = emailOfficialExistsInDbAndUnchanged && (isEmailOfficialEditModeOn && (activePerson.emailOfficialVerified || emailOfficialVerifiedLocally));
+  const showResetPassword = emailOfficialExistsInDbAndUnchanged && (activePerson.emailOfficialVerified || emailOfficialVerifiedLocally);
   const showVerifyEmailLink = emailOfficialExistsInDbAndUnchanged && ((isEmailOfficialEditModeOn && !emailOfficialVerifiedToNotExist) || emailOfficialNotVerifiedInDbOrLocally);
 
   return (
@@ -261,13 +298,6 @@ Please note that we prefer that you keep your Slack account connected to your pe
                   </SpanWithLinkStyle>
                 </ActionOption>
               )}
-              {showResetPassword && (
-                <ActionOption>
-                  <SpanWithLinkStyle onClick={() => resetPassword()}>
-                    Reset password
-                  </SpanWithLinkStyle>
-                </ActionOption>
-              )}
               {showEditEmailLink && (
                 <ActionOption>
                   <SpanWithLinkStyle onClick={() => setIsEmailOfficialEditModeOnLocal(true)}>
@@ -293,6 +323,13 @@ Please note that we prefer that you keep your Slack account connected to your pe
                     </SpanWithLinkStyle>
                   </ActionOption>
                 </>
+              )}
+              {showResetPassword && (
+                <ActionOption>
+                  <SpanWithLinkStyle onClick={() => resetPassword()}>
+                    Reset password
+                  </SpanWithLinkStyle>
+                </ActionOption>
               )}
             </ActionOptionList>
           </ActionOptionContainerLeft8>
@@ -346,6 +383,45 @@ Please note that we prefer that you keep your Slack account connected to your pe
             />
           </div>
         </>
+      )}
+      {showNewPasswordNotification && (
+        <>
+          <ActionOptionContainerOverflow>
+            <ActionOptionContainerLeft8>
+              <ActionOptionList>
+                <ActionOption>
+                  {newPasswordNotificationCopied ? 'Copied!' : (
+                    <CopyToClipboard text={newPasswordNotification} onCopy={() => newPasswordNotificationOnCopy(true)}>
+                      <SpanWithLinkStyle>
+                        Copy to Clipboard
+                      </SpanWithLinkStyle>
+                    </CopyToClipboard>
+                  )}
+                </ActionOption>
+                <ActionOption>
+                  Please send to volunteer via Slack.
+                </ActionOption>
+              </ActionOptionList>
+            </ActionOptionContainerLeft8>
+          </ActionOptionContainerOverflow>
+          <div>
+            <TextField
+              multiline
+              rows={5}
+              value={newPasswordNotification}
+              // onChange={(e) => setEmailText(e.target.value)}
+              placeholder="Enter text to be copied..."
+              fullWidth
+              variant="outlined"
+              margin="normal"
+            />
+          </div>
+        </>
+      )}
+      {emailError.length > 0 && (
+        <Alert icon={<ErrorOutline fontSize="inherit" />} severity="error">
+          {emailError}
+        </Alert>
       )}
       <div>
         <pre id="jsonResults" style={{ marginTop: '11px', fontWeight: '700' }} />
